// src/app/shared/model/page.model.ts
export interface users {
  sNo: number;
  img: string;
  userName: string;
  phone: string;
  email: string;
  role: string;
  createdOn: string;
  status: string;
  isSelected?: boolean;
}